contanct form initiated

+++ post/views.py
@@ -10,7 +10,7 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 
 
 # Create your views here.
-class PostListView(View, LoginRequiredMixin):
+class PostListView(LoginRequiredMixin, View):
     def get(self, request, *args, **kwargs):
         posts = Post.objects.filter(user=request.user)
         paginator = Paginator(posts, 8)
@@ -34,9 +34,11 @@ class PostDetailView(View):
         return render(request, "post/detail", context={"post": post})
 
 
-class PostEditView(View):
+class PostEditView(LoginRequiredMixin, View):
     def get(self, request, id, *args, **kwargs):
         post = Post.objects.get(id=id)
+        if post.user != request.user:
+            return redirect(reverse("404_error"))
         form = PostForm(instance=post)
         return render(request, "post/edit.html", context={"form": form})
 
@@ -49,7 +51,7 @@ class PostEditView(View):
         return render(request, "post/edit.html", context={"form": form})
 
 
-class PostCreateView(View):
+class PostCreateView(LoginRequiredMixin, View):
     def get(self, request, *args, **kwargs):
         form = PostForm()
         return render(request, "post/create.html", context={"form": form})
@@ -64,9 +66,11 @@ class PostCreateView(View):
         return render(request, "post/create.html", context={"form": form})
 
 
-class PostDeleteView(View):
+class PostDeleteView(LoginRequiredMixin, View):
     def post(self, request, *args, **kwargs):
         post = Post.objects.get(id=request.POST.get("id"))
+        if post.user != request.user:
+            return redirect(reverse("404_error"))
         post.delete()
         return redirect(reverse("post-list"))
